// src/payload-types.ts
export interface Resource {
  id: number;
  title?: string | null;
  description?: string | null;
  uri?: string | null;
  tags?:
    | {
        tag?: string | null;
        id?: string | null;
      }[]
    | null;
  type?: ('link' | 'video' | 'image' | 'audio' | 'document') | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "sessions".
 */
export interface Session {
  id: number;
  title?: string | null;
  description?: string | null;
  topics?:
    | {
        topic?: string | null;
        id?: string | null;
      }[]
    | null;
  date?: string | null;
  resources?: (number | Resource)[] | null;
  updatedAt: string;
  createdAt: string;
}